requirements.txt
# Core dependencies
marker-pdf